# api/restaurant/restful/restaurant_restful_swagger.yaml
basePath: /api
definitions:
  github_com_blackhorseya_godine_entity_domain_restaurant_model.Address:
    properties:
      city:
        description: City is the city where the restaurant is located.
        type: string
      state:
        description: State is the state where the restaurant is located.
        type: string
      street:
        description: Street is the street address of the restaurant.
        type: string
      zip_code:
        description: ZipCode is the postal code of the restaurant's location.
        type: string
    type: object
  github_com_blackhorseya_godine_entity_domain_restaurant_model.MenuItem:
    properties:
      description:
        description: Description provides details about the menu item.
        type: string
      id:
        description: ID is the unique identifier of the menu item.
        type: string
      is_available:
        description: IsAvailable indicates whether the menu item is available.
        type: boolean
      name:
        description: Name is the name of the menu item.
        type: string
      price:
        description: Price is the cost of the menu item.
        type: number
    type: object
  github_com_blackhorseya_godine_entity_domain_restaurant_model.Restaurant:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.Address'
        description: Address is the address of the restaurant.
      id:
        description: ID is the unique identifier of the restaurant.
        type: string
      is_open:
        description: IsOpen indicates whether the restaurant is open for business.
        type: boolean
      menu:
        description: Menu is the list of menu items available in the restaurant.
        items:
          $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.MenuItem'
        type: array
      name:
        description: Name is the name of the restaurant.
        type: string
    type: object
  items.PostPayload:
    properties:
      description:
        example: ""
        type: string
      name:
        example: items item name
        type: string
      price:
        example: 10
        type: number
    required:
    - name
    - price
    type: object
  responsex.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  restaurants.PatchWithStatusPayload:
    properties:
      is_open:
        example: true
        type: boolean
    type: object
  restaurants.PostPayload:
    properties:
      description:
        type: string
      name:
        example: restaurant name
        type: string
    required:
    - name
    type: object
info:
  contact:
    email: blackhorseya@gmail.com
    name: Sean Zheng
    url: https://blog.seancheng.space
  description: Godine Restaurant Restful API document.
  license:
    name: GPL-3.0
    url: https://spdx.org/licenses/GPL-3.0-only.html
  title: Godine Restaurant Restful API
  version: 0.1.0
paths:
  /healthz:
    get:
      consumes:
      - application/json
      description: Check the health of the service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsex.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Check the health of the service.
      tags:
      - health
  /v1/restaurants:
    get:
      consumes:
      - application/json
      description: Get the restaurant list.
      parameters:
      - default: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: total count
              type: number
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.Restaurant'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Get the restaurant list.
      tags:
      - restaurants
    post:
      consumes:
      - application/json
      description: Create the restaurant.
      parameters:
      - description: restaurant payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/restaurants.PostPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.Restaurant'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Create the restaurant.
      tags:
      - restaurants
  /v1/restaurants/{restaurant_id}:
    delete:
      consumes:
      - application/json
      description: Delete the restaurant by id.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/responsex.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Delete the restaurant by id.
      tags:
      - restaurants
    get:
      consumes:
      - application/json
      description: Get the restaurant by id.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.Restaurant'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Get the restaurant by id.
      tags:
      - restaurants
    put:
      consumes:
      - application/json
      description: Update the restaurant by id.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: restaurant payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.Restaurant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.Restaurant'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Update the restaurant by id.
      tags:
      - restaurants
  /v1/restaurants/{restaurant_id}/items:
    get:
      consumes:
      - application/json
      description: Get the items list.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      - default: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: total count
              type: number
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.MenuItem'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Get the items list.
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Add a items item.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: items item payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/items.PostPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.MenuItem'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Add a items item.
      tags:
      - items
  /v1/restaurants/{restaurant_id}/items/{item_id}:
    delete:
      consumes:
      - application/json
      description: Delete the items item by id.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: item id
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/responsex.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Delete the items item by id.
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Get the items item by id.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: item id
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.MenuItem'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Get the items item by id.
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update the items item by id.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: item id
        in: path
        name: item_id
        required: true
        type: string
      - description: items item payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.MenuItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.MenuItem'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Update the items item by id.
      tags:
      - items
  /v1/restaurants/{restaurant_id}/status:
    patch:
      consumes:
      - application/json
      description: Update the restaurant status by id.
      parameters:
      - description: restaurant id
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: restaurant status payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/restaurants.PatchWithStatusPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responsex.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_blackhorseya_godine_entity_domain_restaurant_model.Restaurant'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsex.Response'
      summary: Update the restaurant status by id.
      tags:
      - restaurants
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"